#700 - Postgres ARRAY and JSON expressions cast Long to ::integer rather than ::bigint

+++ src/main/java/com/avaje/ebeaninternal/server/type/ScalarTypeArrayList.java
@@ -23,7 +23,7 @@ public class ScalarTypeArrayList extends ScalarTypeJsonCollection<List> {
 
   private static ScalarTypeArrayList UUID = new ScalarTypeArrayList("uuid", DocPropertyType.UUID, ArrayElementConverter.UUID);
 
-  private static ScalarTypeArrayList LONG = new ScalarTypeArrayList("integer", DocPropertyType.LONG, ArrayElementConverter.LONG);
+  private static ScalarTypeArrayList LONG = new ScalarTypeArrayList("bigint", DocPropertyType.LONG, ArrayElementConverter.LONG);
 
   private static ScalarTypeArrayList INTEGER = new ScalarTypeArrayList("integer", DocPropertyType.INTEGER, ArrayElementConverter.INTEGER);
